requirements.txt
streamlit>=1.35.0
numpy>=1.26.0
plotly>=5.20.0
matplotlib>=3.8.0
numba>=0.59.0
scipy>=1.12.0
Pillow>=10.0.0
